realtime_tracking/requirements.txt
fastapi
uvicorn
opencv-python
mediapipe
firebase-admin
jinja2
pillow